backend/requirements.txt
azure-core
azure-storage-blob
fastapi
gunicorn
nltk
numpy
openai
pandas
PyPDF2
scikit-learn
mysql-connector-python
websockets
python-multipart
uvicorn